Events Gallery Section added, responsive layout of prizes section & see you soon message on front page
