+++ package.json
@@ -32,7 +32,7 @@
     "framer-motion": "^11.18.2",
     "fs-extra": "^11.3.0",
     "gsap": "^3.12.7",
-    "next": "^15.2.1",
+    "next": "^15.5.4",
     "nvm": "^0.0.4",
     "postcss": "^8.4.49",
     "react": "^18.3.1",
