+++ src/assets/events_gallery_data/images.js
@@ -1,0 +1,45 @@
+import image1 from "../../assets/images/19.jpg"
+import image2 from "../../assets/images/20.png"
+import image3 from "../../assets/images/21.png"
+import image4 from "../../assets/images/22.jpg"
+import image5 from "../../assets/images/23.png"
+import image6 from "../../assets/images/24.png"
+import image7 from "../../assets/images/25.jpg"
+import image8 from "../../assets/images/26.png"
+import image9 from "../../assets/images/27.png"
+import image10 from "../../assets/images/28.png"
+import image11 from "../../assets/images/29.png"
+import image12 from "../../assets/images/30.jpg"
+import image13 from "../../assets/images/31.png"
+import image14 from "../../assets/images/32.png"
+import image15 from "../../assets/images/33.jpg"
+import image16 from "../../assets/images/34.png"
+import image17 from "../../assets/images/35.png"
+import image18 from "../../assets/images/36.jpg"
+const data =
+{
+        pic1: image1,
+        pic2: image2,
+        pic3: image3,
+        pic4: image4,
+        pic5: image5,
+        pic6: image6,
+        pic7: image7,
+        pic8: image8,
+        pic9: image9,
+        pic10: image10,
+        pic11: image11,
+        pic12: image12,
+        pic13: image13,
+        pic14: image14,
+        pic15: image15,
+        pic16: image16,
+        pic17: image17,
+        pic18: image18,
+
+}
+
+export default data;
+
+
+
