--- src/assets/last_electrothon_data/images.js
@@ -1,45 +1,0 @@
-import image1 from "../../assets/images/1.webp"
-import image2 from "../../assets/images/2.webp"
-import image3 from "../../assets/images/3.webp"
-import image4 from "../../assets/images/4.webp"
-import image5 from "../../assets/images/5.webp"
-import image6 from "../../assets/images/6.webp"
-import image7 from "../../assets/images/7.webp"
-import image8 from "../../assets/images/8.webp"
-import image9 from "../../assets/images/9.webp"
-import image10 from "../../assets/images/10.webp"
-import image11 from "../../assets/images/11.webp"
-import image12 from "../../assets/images/12.webp"
-import image13 from "../../assets/images/13.webp"
-import image14 from "../../assets/images/14.webp"
-import image15 from "../../assets/images/15.webp"
-import image16 from "../../assets/images/16.webp"
-import image17 from "../../assets/images/17.webp"
-import image18 from "../../assets/images/18.webp"
-const data =
-{
-        pic1: image1,
-        pic2: image2,
-        pic3: image3,
-        pic4: image4,
-        pic5: image5,
-        pic6: image6,
-        pic7: image7,
-        pic8: image8,
-        pic9: image9,
-        pic10: image10,
-        pic11: image11,
-        pic12: image12,
-        pic13: image13,
-        pic14: image14,
-        pic15: image15,
-        pic16: image16,
-        pic17: image17,
-        pic18: image18,
-
-}
-
-export default data;
-
-
-
