+++ src/components/events_gallery/slider.module.css
@@ -1,0 +1,126 @@
+.dualSlider {
+    display: flex;
+    flex-direction: column;
+    gap: 40px;
+    margin: 0 auto;
+    padding: 0;
+    justify-content: center;
+    align-items: center;
+    /* margin-bottom: 80px; */
+    width: 100%;
+}
+
+.slider {
+    width: 100%;
+    max-width: 1200px;
+    position: relative;
+}
+
+.imageWrapper {
+    position: relative;
+    overflow: hidden;
+    border-radius: 5px;
+    border: 3px solid #ffffff;
+    max-width: 100%;
+    will-change: transform; 
+}
+
+.image {
+    width: 100%;
+    height: auto;
+    object-fit: cover;
+    transition: transform 0.3s ease-in-out;
+}
+
+.image:hover {
+    transform: scale(1.1);
+}
+
+.swiper-slide {
+    min-height: 200px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: auto;
+    flex-shrink: 0;
+    overflow: hidden
+}
+
+.swiper-button-next,
+.swiper-button-prev {
+    color: gray;
+    background-color: rgba(255, 255, 255, 0.7);
+    border-radius: 50%;
+    padding: 10px;
+    transition: background-color 0.3s ease, color 0.3s ease;
+}
+
+.swiper-button-next:hover,
+.swiper-button-prev:hover {
+    color: #333;
+    background-color: rgba(0, 0, 0, 0.1);
+}
+
+@media (min-width: 1025px) {
+    .slider {
+        max-width: 1200px;
+    }
+
+    .imageWrapper {
+        border-radius: 10px;
+    }
+
+    .image {
+        width: 100%;
+        height: auto;
+    }
+}
+
+
+@media (min-width: 768px) and (max-width: 1024px) {
+    .slider {
+        max-width: 90%;
+    }
+
+    .imageWrapper {
+        border-radius: 8px;
+    }
+
+    .swiper-button-next,
+    .swiper-button-prev {
+        padding: 8px;
+        font-size: 12px;
+    }
+}
+
+
+@media (max-width: 767px) {
+    .dualSlider {
+        gap: 30px;
+        padding: 0 20px;
+    }
+
+    .slider {
+        max-width: 100%;
+        margin: 0 auto;
+    }
+
+    .imageWrapper {
+        border-radius: 6px;
+    }
+
+    .image {
+        width: 100%;
+        height: auto;
+    }
+
+    .swiper-slide {
+        width: 100%;
+    }
+
+    .swiper-button-next,
+    .swiper-button-prev {
+        padding: 6px;
+        font-size: 10px;
+    }
+}
